+++ src/batch_coordinator/simple/mod.rs
@@ -1,3 +1,6 @@
+mod index;
+
+use index::Index;
 use std::{
     collections::HashSet,
     fs::{File, OpenOptions},
@@ -10,7 +13,7 @@ use uuid::Uuid;
 
 use crate::{
     batch_coordinator::BatchInfo, error::RisklessResult,
-    messages::commit_batch_request::CommitBatchRequest, simple_batch_coordinator::index::Index,
+    messages::commit_batch_request::CommitBatchRequest,
 };
 
 use crate::batch_coordinator::{
@@ -264,7 +267,7 @@ impl BatchCoordinator for SimpleBatchCoordinator {
 
     /// No-op as this operation is not supported in the SimpleBatchCoordinator.
     async fn delete_files(&self, _request: DeleteFilesRequest) {}
-    
+
     /// Always returns false.
     async fn is_safe_to_delete_file(&self, _object_key: String) -> bool {
         false
