--- src/simple_batch_coordinator/mod.rs
@@ -1,4 +1,0 @@
-mod index;
-mod simple;
-
-pub use simple::SimpleBatchCoordinator;
